Change LED connector style. WIP.
I guess this is the start of Rev B. I realized that I could
probably standardize the ribbon cable. In theory, I could also create a
PCB for the front panel LEDs, then use a standard 2x5-pin ribbon cable
to connect the two boards.

Damn, that'd be a great idea.