Beeg refactor
Move a lot of shit out of the global scope, into the main function, and
also break out a lot of shit into a class. I tried to move the MQTT
stuff but it got mad when I did that. I'd like to see if that can move,
but not sure if it's a big deal.

I'm sure I violated a ton of Python design patterns because I'm stupid.
Enjoy.